fix-test: Moved all UserPage-tests to one project instead of two
