+++ xUnitTests/UserPageServiceTests.cs
@@ -19,6 +19,68 @@ namespace HealthCareABApi.Tests
             _userPageService = new UserPageService(_userRepositoryMock.Object);
         }
 
+        //GET User Information
+
+        [Fact]
+        public async Task GetUserInformationAsync_ShouldReturnUserDto_WhenUserExists()
+        {
+            // Arrange
+            var userId = "123";
+            var user = new User
+            {
+                Id = userId,
+                Username = "testuser",
+                Roles = new List<string> { "Admin", "User" },
+                FirstName = "John",
+                LastName = "Doe",
+                Email = "john.doe@example.com",
+                Phone = "1234567890",
+                Address = "123 Main St",
+                Gender = "Male",
+                DateOfBirth = new DateTime(1990, 1, 1)
+            };
+
+            _userRepositoryMock.Setup(repo => repo.GetByIdAsync(userId))
+                .ReturnsAsync(user);
+
+            // Act
+            var result = await _userPageService.GetUserInformationAsync(userId);
+
+            // Assert
+            Assert.NotNull(result);
+            Assert.Equal(user.Username, result.Username);
+            Assert.Equal(user.Roles, result.Roles);
+            Assert.Equal(user.FirstName, result.FirstName);
+            Assert.Equal(user.LastName, result.LastName);
+            Assert.Equal(user.Email, result.Email);
+            Assert.Equal(user.Phone, result.Phone);
+            Assert.Equal(user.Address, result.Address);
+            Assert.Equal(user.Gender, result.Gender);
+            Assert.Equal(user.DateOfBirth, result.DateOfBirth);
+
+            _userRepositoryMock.Verify(repo => repo.GetByIdAsync(userId), Times.Once);
+        }
+
+        [Fact]
+        public async Task GetUserInformationAsync_ShouldReturnNull_WhenUserDoesNotExist()
+        {
+            // Arrange
+            var userId = "123";
+
+            _userRepositoryMock.Setup(repo => repo.GetByIdAsync(userId))
+                .ReturnsAsync((User)null);
+
+            // Act
+            var result = await _userPageService.GetUserInformationAsync(userId);
+
+            // Assert
+            Assert.Null(result);
+            _userRepositoryMock.Verify(repo => repo.GetByIdAsync(userId), Times.Once);
+
+        }
+
+
+        // UPDATE User Information
         [Fact]
         public async Task UpdateUserInformationAsync_ShouldUpdateUser_WhenUserExistsAndValidDataProvided()
         {
@@ -94,7 +156,6 @@ namespace HealthCareABApi.Tests
             _userRepositoryMock.Verify(repo => repo.GetByIdAsync(userId), Times.Once);
             _userRepositoryMock.Verify(repo => repo.UpdateAsync(It.IsAny<string>(), It.IsAny<User>()), Times.Never);
 
-            // Testcomment to trigger workflow after changes in YAML-file
         }
     }
 }
